+++ Cargo.toml
@@ -9,7 +9,6 @@ members = [
     "vortex-btrblocks",
     "vortex-buffer",
     "vortex-datafusion",
-    "vortex-datetime-dtype",
     "vortex-dtype",
     "vortex-error",
     "vortex-expr",
@@ -171,7 +170,6 @@ vortex-btrblocks = { version = "0.25.2", path = "./vortex-btrblocks" }
 vortex-buffer = { version = "0.25.2", path = "./vortex-buffer" }
 vortex-bytebool = { version = "0.25.2", path = "./encodings/bytebool" }
 vortex-datafusion = { version = "0.25.2", path = "./vortex-datafusion" }
-vortex-datetime-dtype = { version = "0.25.2", path = "./vortex-datetime-dtype" }
 vortex-datetime-parts = { version = "0.25.2", path = "./encodings/datetime-parts" }
 vortex-dict = { version = "0.25.2", path = "./encodings/dict" }
 vortex-dtype = { version = "0.25.2", path = "./vortex-dtype", default-features = false }
