+++ bench-vortex/src/bin/notimplemented.rs
@@ -10,7 +10,7 @@ use vortex::arrays::{
     VarBinViewArray,
 };
 use vortex::buffer::buffer;
-use vortex::datetime_dtype::{TIME_ID, TemporalMetadata, TimeUnit};
+use vortex::dtype::datetime::{TIME_ID, TemporalMetadata, TimeUnit};
 use vortex::dtype::{DType, ExtDType, Nullability, PType};
 use vortex::encodings::alp::{ALPArray, Exponents, RDEncoder};
 use vortex::encodings::bytebool::ByteBoolArray;
