+++ bench-vortex/src/clickbench.rs
@@ -11,8 +11,8 @@ use futures::{StreamExt, TryStreamExt, stream};
 use tokio::fs::{OpenOptions, create_dir_all};
 use tracing::info;
 use vortex::TryIntoArray;
-use vortex::arrow::FromArrowType;
 use vortex::dtype::DType;
+use vortex::dtype::arrow::FromArrowType;
 use vortex::error::{VortexError, vortex_err};
 use vortex::file::{DEFAULT_REGISTRY, VORTEX_FILE_EXTENSION, VortexWriteOptions};
 use vortex::layout::{LayoutRegistry, LayoutRegistryExt};
