+++ encodings/datetime-parts/Cargo.toml
@@ -21,7 +21,6 @@ getrandom_v02 = { workspace = true }
 rkyv = { workspace = true }
 vortex-array = { workspace = true }
 vortex-buffer = { workspace = true }
-vortex-datetime-dtype = { workspace = true }
 vortex-dtype = { workspace = true }
 vortex-error = { workspace = true }
 vortex-mask = { workspace = true }
