+++ encodings/datetime-parts/src/timestamp.rs
@@ -1,4 +1,4 @@
-use vortex_datetime_dtype::TimeUnit;
+use vortex_dtype::datetime::TimeUnit;
 use vortex_error::{VortexResult, vortex_bail};
 
 pub const SECONDS_PER_DAY: i64 = 86_400; // 24 * 60 * 60
