+++ vortex-array/Cargo.toml
@@ -51,8 +51,7 @@ rkyv = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
 static_assertions = { workspace = true }
 vortex-buffer = { workspace = true, features = ["arrow", "rkyv"] }
-vortex-datetime-dtype = { workspace = true }
-vortex-dtype = { workspace = true, features = ["rkyv", "serde"] }
+vortex-dtype = { workspace = true, features = ["arrow", "rkyv", "serde"] }
 vortex-error = { workspace = true, features = [
     "flatbuffers",
     "flexbuffers",
