+++ vortex-array/src/arrays/datetime/mod.rs
@@ -3,7 +3,7 @@ mod test;
 
 use std::sync::Arc;
 
-use vortex_datetime_dtype::{DATE_ID, TIME_ID, TIMESTAMP_ID, TemporalMetadata, TimeUnit};
+use vortex_dtype::datetime::{DATE_ID, TIME_ID, TIMESTAMP_ID, TemporalMetadata, TimeUnit};
 use vortex_dtype::{DType, ExtDType};
 use vortex_error::{VortexError, vortex_err, vortex_panic};
