+++ vortex-array/src/arrow/mod.rs
@@ -4,8 +4,6 @@ use arrow_array::ArrayRef;
 use arrow_schema::DataType;
 use vortex_error::VortexResult;
 
-pub use crate::arrow::dtype::{infer_data_type, infer_schema};
-
 mod array;
 mod datum;
 mod dtype;
@@ -14,14 +12,6 @@ pub use datum::*;
 
 pub trait FromArrowArray<A> {
     fn from_arrow(array: A, nullable: bool) -> Self;
-}
-
-pub trait FromArrowType<T>: Sized {
-    fn from_arrow(value: T) -> Self;
-}
-
-pub trait TryFromArrowType<T>: Sized {
-    fn try_from_arrow(value: T) -> VortexResult<Self>;
 }
 
 pub trait IntoArrowArray {
