+++ vortex-datafusion/src/persistent/sink.rs
@@ -12,9 +12,9 @@ use datafusion_physical_plan::{DisplayAs, DisplayFormatType};
 use futures::{StreamExt, TryStreamExt};
 use rand::distr::{Alphanumeric, SampleString};
 use vortex_array::TryIntoArray;
-use vortex_array::arrow::FromArrowType;
 use vortex_array::stream::ArrayStreamAdapter;
 use vortex_dtype::DType;
+use vortex_dtype::arrow::FromArrowType;
 use vortex_error::VortexError;
 use vortex_file::{VORTEX_FILE_EXTENSION, VortexWriteOptions};
 use vortex_io::{ObjectStoreWriter, VortexWrite};
