--- vortex-datetime-dtype/Cargo.toml
@@ -1,30 +1,0 @@
-[package]
-name = "vortex-datetime-dtype"
-version = { workspace = true }
-description = "Vortex datetime extension dtype"
-homepage = { workspace = true }
-repository = { workspace = true }
-authors = { workspace = true }
-license = { workspace = true }
-keywords = { workspace = true }
-include = { workspace = true }
-edition = { workspace = true }
-rust-version = { workspace = true }
-categories = { workspace = true }
-readme = { workspace = true }
-
-[dependencies]
-arrow-schema = { workspace = true, optional = true }
-jiff = { workspace = true }
-num_enum = { workspace = true }
-serde = { workspace = true, features = ["derive"] }
-vortex-dtype = { workspace = true, features = ["serde"] }
-vortex-error = { workspace = true }
-
-
-[features]
-default = ["arrow"]
-arrow = ["dep:arrow-schema"]
-
-[lints]
-workspace = true
