+++ vortex-dtype/src/datetime/mod.rs
@@ -1,4 +1,7 @@
+//! Module for Vortex datetime extension types.
+#![cfg(feature = "arrow")]
 pub mod arrow;
+
 mod temporal;
 mod unit;
