+++ vortex-scalar/src/arrow.rs
@@ -1,7 +1,7 @@
 use std::sync::Arc;
 
 use arrow_array::*;
-use vortex_datetime_dtype::{TemporalMetadata, TimeUnit, is_temporal_ext_type};
+use vortex_dtype::datetime::{TemporalMetadata, TimeUnit, is_temporal_ext_type};
 use vortex_dtype::{DType, PType};
 use vortex_error::{VortexError, vortex_bail, vortex_err};
