+++ vortex-scalar/src/datafusion.rs
@@ -4,8 +4,8 @@ use std::sync::Arc;
 
 use datafusion_common::ScalarValue;
 use vortex_buffer::ByteBuffer;
-use vortex_datetime_dtype::arrow::make_temporal_ext_dtype;
-use vortex_datetime_dtype::{TemporalMetadata, TimeUnit, is_temporal_ext_type};
+use vortex_dtype::datetime::arrow::make_temporal_ext_dtype;
+use vortex_dtype::datetime::{TemporalMetadata, TimeUnit, is_temporal_ext_type};
 use vortex_dtype::half::f16;
 use vortex_dtype::{DType, Nullability, PType};
 use vortex_error::VortexError;
