+++ vortex/src/lib.rs
@@ -7,10 +7,10 @@ pub use vortex_file as file;
 #[cfg(feature = "files")]
 pub use vortex_io as io;
 pub use {
-    vortex_btrblocks as compressor, vortex_buffer as buffer,
-    vortex_datetime_dtype as datetime_dtype, vortex_dtype as dtype, vortex_error as error,
-    vortex_expr as expr, vortex_flatbuffers as flatbuffers, vortex_ipc as ipc,
-    vortex_layout as layout, vortex_mask as mask, vortex_proto as proto, vortex_scalar as scalar,
+    vortex_btrblocks as compressor, vortex_buffer as buffer, vortex_dtype as dtype,
+    vortex_error as error, vortex_expr as expr, vortex_flatbuffers as flatbuffers,
+    vortex_ipc as ipc, vortex_layout as layout, vortex_mask as mask, vortex_proto as proto,
+    vortex_scalar as scalar,
 };
 
 pub mod encodings {
